+++ grafana_import/constants.py
@@ -1,3 +1,3 @@
 PKG_NAME = 'grafana-import'
-PKG_VERSION = '0.0.2'
+PKG_VERSION = '0.1.0'
 CONFIG_NAME = 'conf/grafana-import.yml'