+++ rebuild.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
-rm -rf /usr/local/lib/python3.6/site-packages/grafana_imprt /usr/local/lib/python3.6/site-packages/grafana_import-0.0.1-py3.6.egg-info/
+rm -rf /usr/local/lib/python3.6/site-packages/grafana_import /usr/local/lib/python3.6/site-packages/grafana_import-0.0.1-py3.6.egg-info/
 
 pip3 install  .