+++ HISTORY.rst
@@ -3,9 +3,24 @@
 History
 =======
 
+# 0.1.0 (2022-02-01)
+====================
+
+* fixe behavior for dashboard moved from one folder to another
+
+# 0.0.3 (2022-02-31)
+====================
+
+* add remove dashboard feature
+
+# 0.0.2 (2022-01-07)
+====================
+
+* change config file format from json to yml
+* add labels in config to define multi grafana servers.
+
 # 0.0.1 (2021-03-15)
 ====================
 
 * First release on github.
-